# src/main/java/com/theawesomeengineer/taskmanager/api/TasksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.16.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.theawesomeengineer.taskmanager.api;

import com.theawesomeengineer.taskmanager.model.Error;
import com.theawesomeengineer.taskmanager.model.Task;
import com.theawesomeengineer.taskmanager.model.TaskRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-28T12:34:46.383096376+07:00[Asia/Jakarta]", comments = "Generator version: 7.16.0")
@Validated
@Tag(name = "Tasks", description = "the Tasks API")
public interface TasksApi {

    default TasksApiDelegate getDelegate() {
        return new TasksApiDelegate() {};
    }

    public static final String PATH_CREATE_TASK = "/tasks";
    /**
     * POST /tasks : Create a new task
     * Create a new task with the provided information
     *
     * @param taskRequest  (required)
     * @return Task created successfully (status code 201)
     *         or Invalid input (status code 400)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "createTask",
        summary = "Create a new task",
        description = "Create a new task with the provided information",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Task created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Task.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = TasksApi.PATH_CREATE_TASK,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Task> createTask(
        @Parameter(name = "TaskRequest", description = "", required = true) @Valid @RequestBody TaskRequest taskRequest
    ) {
        return getDelegate().createTask(taskRequest);
    }


    public static final String PATH_DELETE_TASK = "/tasks/{id}";
    /**
     * DELETE /tasks/{id} : Delete task by ID
     * Delete a specific task by its ID
     *
     * @param id Task ID (required)
     * @return Task deleted successfully (status code 204)
     *         or Task not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "deleteTask",
        summary = "Delete task by ID",
        description = "Delete a specific task by its ID",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Task deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Task not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = TasksApi.PATH_DELETE_TASK,
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> deleteTask(
        @NotNull @Parameter(name = "id", description = "Task ID", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return getDelegate().deleteTask(id);
    }


    public static final String PATH_GET_ALL_TASKS = "/tasks";
    /**
     * GET /tasks : Get all tasks
     * Retrieve a list of all tasks
     *
     * @return List of tasks retrieved successfully (status code 200)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getAllTasks",
        summary = "Get all tasks",
        description = "Retrieve a list of all tasks",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of tasks retrieved successfully", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Task.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TasksApi.PATH_GET_ALL_TASKS,
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Task>> getAllTasks(
        
    ) {
        return getDelegate().getAllTasks();
    }


    public static final String PATH_GET_TASK_BY_ID = "/tasks/{id}";
    /**
     * GET /tasks/{id} : Get task by ID
     * Retrieve a specific task by its ID
     *
     * @param id Task ID (required)
     * @return Task retrieved successfully (status code 200)
     *         or Task not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getTaskById",
        summary = "Get task by ID",
        description = "Retrieve a specific task by its ID",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Task retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Task.class))
            }),
            @ApiResponse(responseCode = "404", description = "Task not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TasksApi.PATH_GET_TASK_BY_ID,
        produces = { "application/json" }
    )
    
    default ResponseEntity<Task> getTaskById(
        @NotNull @Parameter(name = "id", description = "Task ID", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return getDelegate().getTaskById(id);
    }


    public static final String PATH_UPDATE_TASK = "/tasks/{id}";
    /**
     * PUT /tasks/{id} : Update task by ID
     * Update an existing task with the provided information
     *
     * @param id Task ID (required)
     * @param taskRequest  (required)
     * @return Task updated successfully (status code 200)
     *         or Invalid input (status code 400)
     *         or Task not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "updateTask",
        summary = "Update task by ID",
        description = "Update an existing task with the provided information",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Task updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Task.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Task not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = TasksApi.PATH_UPDATE_TASK,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Task> updateTask(
        @NotNull @Parameter(name = "id", description = "Task ID", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "TaskRequest", description = "", required = true) @Valid @RequestBody TaskRequest taskRequest
    ) {
        return getDelegate().updateTask(id, taskRequest);
    }

}